+++ UnitTests/MockURLOpener.swift
@@ -1,0 +1,11 @@
+import Foundation
+@testable import PopupBridge
+
+class MockURLOpener: URLOpener {
+
+    var venmoInstalled = false
+
+    func isVenmoAppInstalled() -> Bool {
+        venmoInstalled
+    }
+}